Add a two-argument `shortest_timespan_containing` method
It's not uncommon to want merge exactly two `TimeSpan`s, but currently
you have to wrap them in an interator in order to pass them to
`shortest_timespan_containing`. This commit adds a method to the
function that accepts two `TimeSpan`s and otherwise does what it says on
the tin.